requirements.txt
asgiref==3.5.0
Django==4.0.2
djangorestframework==3.13.1
drf-writable-nested==0.6.3
pytz==2021.3
sqlparse==0.4.2
tzdata==2021.5
